Adds kata Makefile step call OpenTitan regtool
To write s3L4 for OpenTitan matcha, we want to depend on generated C
header files for definitions of MMIO sizes, offsets, masks, bit
meanings, etc.

To make these header files available to the seL4 build, this change adds
a make prerequisite for the kata build to run regtool.py with output
going to out/kata/opentitan-gen.

In CAmkES CMake files, the INCLUDE attribute can be used to add to the
include path, and this change also does that for the UartDriver
component. (The implementation will be left for a later change, since an
initial straighforwrd attempt saw writes still going to the 16550.)

Change-Id: I4b1f6b5ed7f77e9218e1f8b249550d5ec387f912
GitOrigin-RevId: 9c2792fe0db3906225e2b9e83cb07408cfd7d383