build:iree: Fix the linker script error
riscv_iree.cmake is called twice in try_compile, and adding the linker
script there causes the linker error. Some of the linker flags are
dropped as a result.

Move the executable linker flags to iree.mk CLI to avoid the linker
error.

Add memory usage printout so we can exam the TCM usage.

Also set the ITCM length to true 256K.

Change-Id: I292c44d2f7e23db91feea9145e62fc7dfda3fef0
GitOrigin-RevId: 162834f4983e42ea4b52b603dd162c0d8d9dc72c