build: remove some more executable linker flags
Wired up springbok intrinsics
directly to the executables avoiding the linker flags here.  This cleans
up the now unnecessary linker flags. The only flags in iree.mk are for
dev only and can be removed without breaking the build.

Remove springbog_iree target since it is not used. Move the cmake config
target to a non-phony target so the target would not be rebuilt
unnecessarily.

Change-Id: I8c0a80aa61b23f8f4b92f6aad9998f0da8f55484
GitOrigin-RevId: fac374b919de18980601a7b0bb0de489f68ab2ab