build:riscv_toolchain: Build 11.1.0 rv32imac gcc toolchain tarball
The vector core uses clang now, so we can fastforward the toolchain for
the Security core and SMC to a more recent GCC release. Update the
build script to build for rv32imac architecture. The tarball of
this toolchain is uploaded to x20.

The target of toolchain_vp (GCC RVV) still exists, but it should not used in the
typical cases. It is for debug purpose only, and the cached tarball of
it will not be downloaded via `m tools`

Change-Id: Id4e6f21a19a63f3d8815302e26f6a5dc9c122826
GitOrigin-RevId: fc41d6142b9f5e8f80547dff5551f739eb63eb71