build: Adds a new `hmm` function to help out with developer UX
Something we seem to always need is the ability to figure out which
targets are needed to build something, and what targets do, exactly.
This adds a simple function to setup.sh to give us the ability to
get documentation on various user-facing make targets.

Documentation is in the form of doc-comments using the familiar C
or C++ syntax:

  ## Short summary line
  #
  # Additional description
  target:

The additional `helpmemake.awk` script walks through the whole make
build system and extracts documentation for targets we expect
end-developers to run.

Change-Id: Ibccb6367eae6717268729e2f4b457efa116b69f8
GitOrigin-RevId: 1310757d79cbd75b7a1693bbfa325b0e13375039