build: renode: Update renode make rule to specify the output path
Following the change in
github.com/renode/renode/commit/587f3e93fbccaa398905ee2eef194476d5ee29f8
so the binaries and scripts are built out of the source tree.
This removes the copying of the artifacts at the end of `m renode`, and
ensures the error code check works properly after building renode.

Append `renode_clean` to the `clean` target for a clean
build of renode after `m clean`

Change-Id: Iae918529cf46b452841adc2cf0a3e9b04562e2c9
GitOrigin-RevId: 5b38e6abb59b63219e8841d2f3f9c98f77cff689