build:renode: Allow running multiple renode Sparrow simulation sessions
Define an environment variable RENODE_PORT of 1234 (which is already
preset within Renode). To run a different session concurrently,
change the port to a different value (assuming >= 1234)

Example:
export RENODE_PORT=1235

Renode console will be accessible at localhost:1235, while the GDB
server is at :3334 (3333 + RENODE_PORT - 1234)

./scripts/kshell.sh also accepts the port as the input argument and
access the corresponding emulated pty terminal. If no argument is
set, it assumes the port of 1234.

The feature is a no-op for typical users. If you don't set anything,
the simulator behavior and operation are exactly like before.

Change-Id: Ia9d3595ac9be3fa7f878ba0d29b9dd290da27bfe
GitOrigin-RevId: 9a697557fe9c4e954136dabf38bbd8f93f6080d0