hmm: Use gawk and add completion
We really wanted command line completion for both m and hmm, so this
change migrates helpmemake to use gawk and removes most of the
wrapping UX from the original script and into setup.sh.

To allow for completion, it also adds a new completion function and
registers it for both m and hmm.

In addition, this helps to remove some of the delay that occurs when
using the built-in bash completion.

Change-Id: I3a92c0c9e9a147f342ed135e600c690bbdef7876
GitOrigin-RevId: 34ac7d5d1757b57e5755060756dff934e6491d5c